Permit sequences to be skipped.
Redis does not guarantee delivery of keyspace notifications, hence we cannot insist
that we observe every sequence key using pubsub alone (without some sort of queuing
and acknowledgment mechanism: http://stackoverflow.com/questions/23675394/redis-publish-subscribe-is-redis-guaranteed-to-deliver-the-message-even-under-m#comment36385109_23681531).

However, notifications will be delivered in order, so I believe it is still semantically correct
to say that we have "caught up"–rather than throwing an error–if we observe a sequence
that is greater than the latest we've seen.